test: XXX debugging tests that fail because of destroy/end/data interactions
test/parallel/test-tls-session-with-end-13.js is a minimal repro of bug
sufaced in (and temporarily worked around) test-tls-ticket-cluster.js

test/parallel/test-tls-ticket.js is failing fails because .end() doesn't
cause DoShutdown(), probably its waiting for the previous write('x', cb),
which never callbacks.

test/parallel/test-tls-net-socket-keepalive.js is same as above, but the
test is simpler

Debugging the write problem is what led me to my hack to
writeGeneric()...  that's the wrong way, but what is the right way?