WIP - don't call WriteWrap->Done() while DoWrite() has not returned
What appears to be happening is that `SSL_write()` is writing the 5
bytes of `hello`, but it isn't being flushed out as encrypted text into
the BIO. This might be because its in the middle of SSL code where the
info callback is called, but its about to send the NewTickets, so the
data is just buffered briefly before encrypting it out in an appdata
record. Just a guess.

Since there is nothing to flush, the code wanders down to EncOut(),
where it sees there is nothing to write, so it calls InvokeQueued, which
does a sync w->Done(), in violation of the DoWrite() contract.

Lots more passes, no more failures. Looks very promising.