WIP when should WriteWrap->Done() be called?
This fixes dozens of failing test cases, but I have a few questions
about how the streaming layer is expected to work before committing to
this specific fix.

What appears to be happening is that `SSL_write()` is writing the 5
bytes of `hello`, but it isn't being flushed out as encrypted text into
the BIO. This might be because its in the middle of SSL code where the
info callback is called, but its about to send the NewTickets, so the
data is just buffered briefly before encrypting it out in an appdata
record. Just a guess.

Since there is nothing to flush, the code wanders down to EncOut(),
where it sees there is nothing to write, so it calls InvokeQueued, which
does a sync w->Done(), in violation of the (undocumented) DoWrite() contract.

XXX There is an open question of exactly when to call it. Sync is
illegal, SetImmediate() works, but it seems to me it shouldn't be called
until the data that went into SSL_write() comes out of EncOut() and into
the next stream, and the next stream callbacks, that Done() should
happen.  Tried to make the latter change, but it makes no difference to
the unit tests, so I've no idea if its right or not.