XXX write before destroy can fail after-the-fact
This is a bug, it isn't supposed to happen, but how to prevent it?

When TLS finishes the DoWrite the underlying stream is closed, so the
stream_resource()->Write() in EncOut() fails with write-after-destroy.

What is happening is that with 1.2, SSL_write() from inside the
HANDSHAKE_DONE info callback gets written to the enc_out_ bio
immediately, but with 1.3 it doesn't happen until just AFTER the info
callback returns. This means that the tlsSock.destroy() call Socket._destroy(),
which calls tls_wrap.TLSWrap.prototype.close(), which calls the
underlying stream's (aka "parentWrap") .destroy() ... all before the bio
is filled and the write to the underlying stream occurs.

By the time it occurs, since the underlying stream is destroyed, the
write errors with write-after-destroy.

I'm not sure how this should be fixed. It seems to me that the problem
is that the TLSWrap C++ object should have .destroy() called directly on
it, and it should then forward that .destroy() to the underlying stream
when it's ready, but I don't know the streams code enough to know if
that is possible, advisable, or perhaps just wrong-headed.

See test/parallel/test-tls-destroy-stream.js, and its inline comments.