XXX remove workaround for close_notify before NewSessionTicket
Asked openssl list, they don't think SSL_write() or SSL_shutdown()
should be called in the info callback, but there isn't any way around it
for now, also, it basically works, except for this one edge case.

This only happens with servers that .end() syncly with no data in the
secureConnection event, and that is obscure enough I think we can
accept that those servers don't get ticket tickets delivered.

I intend to drop this revert and go back to the workaround.