XXX remove workaround for close_notify before NewSessionTicket
Code passes with the work-around, but it's a bug.  Not sure how to fix.

Calling SSL_read after after a close_notify is a bit dicy, but might
have to do that, but make sure that only 'session' events occur (ie.,
discard data and anything else). Seems subtly wrong.

Have asked openssl list how this is supposed to work.

This only happens with servers that .end() sync with no data in the
secureConnection event, and maybe those servers don't deserve ticket
support? Not sure.